Move tests for edge-related triggers to their own module
