+++ tests/test_cases/test_cocotb/Makefile
@@ -40,6 +40,7 @@ MODULE := "\
 	test_timing_triggers,\
 	test_scheduler,\
 	test_clock,\
+	test_edge_triggers,\
 	"
 
 ifeq ($(shell python -c "import sys; print(sys.version_info >= (3, 6))"), "True")
